Moved generated API classes to subpackage 'api', to conform to DropWizard conventions.
